rename & cleanup views after nav changes
