+++ src/constants/screens.js
@@ -1,6 +1,5 @@
 export const MAIN_SCREEN_CONTAINER = 'MainScreenContainer'
 export const CONNECTION_SCREEN = 'ConnectionScreen'
-export const HOME_SCREEN = 'HomeScreen'
 export const BROWSE_SONGS_SCREEN = 'BrowseSongsScreen'
 export const SEARCH_SONGS_SCREEN = 'SearchSongsScreen'
 export const PLAYLIST_SCREEN = 'PlaylistScreen'
